backend/requirements.txt
async-generator==1.10
atomicwrites==1.4.1
attrs==21.4.0
bcrypt==3.2.2
behave==1.2.6
cachelib==0.9.0
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
colorama==0.4.5
coverage==6.4.1
cryptography==37.0.4
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
Flask-Session==0.4.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
psycopg==3.0.15
psycopg-binary==3.0.15
psycopg-pool==3.1.1
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-mock==3.8.2
python-dotenv==0.20.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.10
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
